requirements.txt
streamlit==1.31.0
numpy>=1.21.0
pillow>=9.0.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
protobuf==4.25.1 